Update on "[WIP] faketensor cpp: llama3 full vllm command working"
[ghstack-poisoned]
