+++ torch/_dynamo/convert_frame.py
@@ -852,8 +852,7 @@ def trace_frame(
     def run_tracer() -> None:
         try:
             tracer.output.mark_bytecode_tracing_start()
-            cpp_fake = torch._dynamo.config.use_cpp_fake_tensor
-            if cpp_fake:
+            if torch._dynamo.config.use_cpp_fake_tensor:
                 torch._C._activate_cpp_fake_tensor_mode()
             with tracing(tracer.output.tracing_context), tracer.set_current_tx():
                 tracer.run()
@@ -871,7 +870,7 @@ def trace_frame(
                 bisect(tracer.output.shape_env)
             raise
         finally:
-            if cpp_fake:
+            if torch._dynamo.config.use_cpp_fake_tensor:
                 torch._C._deactivate_cpp_fake_tensor_mode()
             tracer.output.call_cleanup_hooks()
             tracer.f_locals = {}
