+++ torch/_dynamo/variables/builder.py
@@ -4120,7 +4120,7 @@ def _wrap_to_fake_tensor_and_record_impl(
             assert tx.fake_mode is not None
             if (
                 config.use_cpp_fake_tensor
-                and torch._C._is_cpp_fake_tensor_mode_active()
+                and torch._C._does_cpp_fake_tensor_mode_exist()
             ):
                 log.debug(
                     "wrap_to_fake (C++ mode) %s %s",
