+++ torch/_functorch/_aot_autograd/collect_metadata_analysis.py
@@ -219,7 +219,7 @@ def run_functionalized_fw_and_collect_metadata(
             # precondition: The passed in function already handles unflattening inputs + flattening outputs
             flat_f_args = pytree.tree_map(_to_fun, flat_args)
             flat_f_args_descs = flat_args_descs
-            _cpp_fake = torch._C._is_cpp_fake_tensor_mode_active()
+            _cpp_fake = torch._C._does_cpp_fake_tensor_mode_exist()
             if _cpp_fake:
                 torch._C._activate_cpp_fake_tensor_mode()
             try:
