+++ torch/_functorch/aot_autograd.py
@@ -532,7 +532,7 @@ def create_aot_state(
     # that we generate in torch.compile.
     stack.enter_context(torch.autograd.set_multithreading_enabled(False))
     stack.enter_context(preserve_rng_state())
-    if not torch._C._is_cpp_fake_tensor_mode_active():
+    if not torch._C._does_cpp_fake_tensor_mode_exist():
         stack.enter_context(fake_mode)
     stack.enter_context(python_dispatcher_mode)
     stack.enter_context(PhiloxStateTracker())
