+++ torch/_inductor/compile_fx.py
@@ -89,6 +89,7 @@ from torch._inductor.utils import (
     get_static_bw_input_idxs,
     InputType,
     is_gpu,
+    maybe_cpp_fake_mode_ctx,
     should_assume_input_aligned,
     should_use_remote_fx_graph_cache,
     tensor_is_aligned,
@@ -708,35 +709,6 @@ def maybe_disable_graph_partition(
         return config.patch(graph_partition=False)
     else:
         return contextlib.nullcontext()
-
-
-@contextlib.contextmanager
-def _cpp_fake_mode_ctx():
-    """Enable the C++ Fake dispatch key in TLS.
-
-    Use this instead of ``with fake_mode:`` when C++ fake tensor mode is
-    active.  It enables the Fake dispatch key so that the C++ fake-tensor
-    kernel handles operations, without ever entering a Python FakeTensorMode.
-    """
-    torch._C._activate_cpp_fake_tensor_mode()
-    try:
-        yield
-    finally:
-        torch._C._deactivate_cpp_fake_tensor_mode()
-
-
-def maybe_cpp_fake_mode_ctx(
-    fake_mode: torch._subclasses.FakeTensorMode,
-) -> AbstractContextManager[None]:
-    """Return the right fake-mode context manager.
-
-    When C++ fake tensor mode is active, returns ``_cpp_fake_mode_ctx()``
-    (enables the Fake dispatch key).  Otherwise returns the Python
-    ``fake_mode`` so it is entered on the TorchDispatchMode stack as usual.
-    """
-    if torch._C._is_cpp_fake_tensor_mode_active():
-        return _cpp_fake_mode_ctx()
-    return fake_mode
 
 
 def cpp_fake_mode_shape_env():
