+++ torch/_inductor/debug.py
@@ -24,6 +24,7 @@ from torch import fx
 from torch._dynamo.repro.after_aot import save_graph_repro
 from torch._dynamo.utils import get_debug_dir
 from torch._inductor import utils
+from torch._inductor.utils import maybe_cpp_fake_mode_ctx
 from torch._logging import getArtifactLogger
 from torch._logging._internal import trace_structured
 from torch._utils_internal import signpost_event
@@ -1255,8 +1256,6 @@ def load_args_and_run_compile_fx_inner(path: str) -> Any:
             return x
 
     fake_mode = torch._subclasses.FakeTensorMode(allow_non_fake_inputs=True)
-    from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
-
     with maybe_cpp_fake_mode_ctx(fake_mode), config.patch("save_args", False):
         args, kwargs = tree_map(handle_tensor, (args, kwargs))
         return compile_fx_inner(*args, **kwargs)
