+++ torch/_inductor/fx_utils.py
@@ -10,13 +10,14 @@ import sympy
 import torch
 import torch.fx
 from torch._dispatch.python import enable_python_dispatcher
+from torch._inductor.utils import maybe_cpp_fake_mode_ctx
 from torch._subclasses.fake_tensor import FakeTensorMode
 
 
 def _get_shape_env():
     from torch._inductor.virtualized import V
 
-    if torch._C._is_cpp_fake_tensor_mode_active():
+    if torch._C._does_cpp_fake_tensor_mode_exist():
         return torch._C._get_cpp_fake_mode_shape_env()
     return V.fake_mode.shape_env
 from torch.fx.experimental.symbolic_shapes import (
@@ -169,8 +170,6 @@ class FakeTensorUpdater:
                     is_valid, args, kwargs = get_fake_args_kwargs(user)
                     if not is_valid:
                         return True
-                    from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
-
                     with (
                         maybe_cpp_fake_mode_ctx(V.fake_mode),
                         enable_python_dispatcher(),
@@ -234,8 +233,6 @@ class FakeTensorUpdater:
             is_valid, args, kwargs = get_fake_args_kwargs(node)
             if not is_valid:
                 continue
-            from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
-
             with maybe_cpp_fake_mode_ctx(V.fake_mode), enable_python_dispatcher():
                 new_fake_tensor = node.target(*args, **kwargs)
