+++ torch/_inductor/ir.py
@@ -108,6 +108,7 @@ from .utils import (
     ir_dataclass,
     is_dynamic,
     is_gpu,
+    maybe_cpp_fake_mode_ctx,
     sympy_dot,
     sympy_index_symbol,
     sympy_index_symbol_with_prefix,
@@ -3907,8 +3908,6 @@ class Layout(OutputSpec):
         return self.device
 
     def get_example(self) -> torch.Tensor:
-        from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
-
         with maybe_cpp_fake_mode_ctx(V.fake_mode):
             return torch.empty_strided(
                 convert_shape_to_symint(self.size),
