+++ torch/_inductor/pattern_matcher.py
@@ -77,6 +77,7 @@ from .._subclasses import FakeTensor, FakeTensorMode
 from ..fx import Transformer
 from . import config
 from .decomposition import select_decomp_table
+from .utils import maybe_cpp_fake_mode_ctx
 from .lowering import fallback_node_due_to_unsupported_type
 
 
@@ -278,10 +279,9 @@ class Match:
                 replacement graph.
 
         """
-        from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
         from torch._inductor.virtualized import NullHandler, V
 
-        if torch._C._is_cpp_fake_tensor_mode_active():
+        if torch._C._does_cpp_fake_tensor_mode_exist():
             context = maybe_cpp_fake_mode_ctx(V.fake_mode)
         elif not isinstance(V.fake_mode, NullHandler) or (V.fake_mode is None):
             context = V.fake_mode
@@ -1535,8 +1535,6 @@ def register_replacement(
         sym_args: list[torch.SymInt] = []
         fake_mode = torch._dynamo.utils.detect_fake_mode(args)
         assert fake_mode is not None
-        from torch._inductor.compile_fx import maybe_cpp_fake_mode_ctx
-
         with maybe_cpp_fake_mode_ctx(fake_mode):
             for i, grad in enumerate(requires_grad):
                 if isinstance(args[i], torch.Tensor):
