+++ torch/fx/experimental/proxy_tensor.py
@@ -682,7 +682,7 @@ def extract_val(val: _ExtractValType, include_real: bool = False) -> _ExtractVal
         return {k: extract_val(v) for k, v in val.items()}
     elif isinstance(val, Tensor):
         if not val.is_sparse:
-            if _has_cpp_fake_tensor and torch._C._is_cpp_fake_tensor_mode_active():
+            if _has_cpp_fake_tensor and torch._C._does_cpp_fake_tensor_mode_exist():
                 return torch.empty_strided(  # revist this
                     val.shape, val.stride(), device=val.device, dtype=val.dtype
                 )
@@ -2728,7 +2728,7 @@ class _MakefxTracer:
             stack.enter_context(decompose(self.decomposition_table))
             if self.fake_tensor_mode and not (
                 _has_cpp_fake_tensor
-                and torch._C._is_cpp_fake_tensor_mode_active()
+                and torch._C._does_cpp_fake_tensor_mode_exist()
             ):
                 stack.enter_context(self.fake_tensor_mode)
             stack.enter_context(self.python_dispatcher_mode)
